Update model inference parameters

+++ src/mindor/dsl/schema/action/impl/model/impl/text_generation.py
@@ -8,7 +8,7 @@ class TextGenerationParamsConfig(BaseModel):
     num_return_sequences: Union[int, str] = Field(default=1, description="The number of generated sequences to return.")
     temperature: Union[float, str] = Field(default=1.0, description="Sampling temperature; higher values produce more random results.")
     top_k: Union[int, str] = Field(default=50, description="Top-K sampling; restricts sampling to the top K tokens.")
-    top_p: Union[int, str] = Field(default=0.9, description="Top-p (nucleus) sampling; restricts sampling to tokens with cumulative probability >= top_p.")
+    top_p: Union[float, str] = Field(default=0.9, description="Top-p (nucleus) sampling; restricts sampling to tokens with cumulative probability >= top_p.")
     batch_size: Union[int, str] = Field(default=32, description="Number of input texts to process in a single batch.")
 
 class TextGenerationModelActionConfig(CommonModelInferenceActionConfig):
